programs/requirements.txt
websockets>=12.0
watchdog>=4.0.0
python-dotenv>=1.0.0
zstandard>=0.22.0
boto3>=1.34.0